Material Rule API -> Surface Rule API
