--- library/worldgen/material_rule/build.gradle
@@ -1,18 +1,0 @@
-plugins {
-	id("qsl.module")
-}
-
-qslModule {
-	name = "Quilt Material Rule API"
-	moduleName = "material_rule"
-	id = "quilt_material_rule"
-	library = "worldgen"
-	description = "An API for managing custom Material Rules."
-	moduleDependencies {
-		core {
-			api("qsl_base")
-			testmodOnly("resource_loader")
-		}
-	}
-	accessWidener()
-}
