--- library/worldgen/material_rule/src/main/java/org/quiltmc/qsl/worldgen/material_rule/mixin/BootstrapMixin.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2022 QuiltMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.quiltmc.qsl.worldgen.material_rule.mixin;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-import net.minecraft.Bootstrap;
-
-import org.quiltmc.qsl.worldgen.material_rule.impl.VanillaSurfaceMaterialRuleTracker;
-
-@Mixin(Bootstrap.class)
-public class BootstrapMixin {
-	@Inject(method = "initialize", at = @At(value = "INVOKE", target = "Lnet/minecraft/util/registry/Registry;freezeBuiltins()V", shift = At.Shift.AFTER))
-	private static void onInitialize(CallbackInfo ci) {
-		VanillaSurfaceMaterialRuleTracker.OVERWORLD.init();
-		VanillaSurfaceMaterialRuleTracker.NETHER.init();
-		VanillaSurfaceMaterialRuleTracker.THE_END.init();
-	}
-}
