--- library/worldgen/material_rule/src/main/resources/quilt_material_rule.mixins.json
@@ -1,13 +1,0 @@
-{
-  "required": true,
-  "package": "org.quiltmc.qsl.worldgen.material_rule.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "BootstrapMixin",
-    "SequenceMaterialRuleMixin",
-    "VanillaSurfaceRulesMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
