--- library/worldgen/material_rule/src/testmod/resources/quilt.mod.json
@@ -1,23 +1,0 @@
-{
-  "schema_version": 1,
-  "quilt_loader": {
-    "group": "org.quiltmc.qsl.worldgen",
-    "id": "quilt_material_rule_testmod",
-    "version": "1.0.0",
-    "metadata": {
-      "name": "Quilt Material Rule API Test Mod",
-      "license": "Apache-2.0"
-    },
-    "intermediate_mappings": "net.fabricmc:intermediary",
-    "load_type": "always",
-    "entrypoints": {
-      "events": [
-        "org.quiltmc.qsl.worldgen.material_rule.QuiltMaterialRuleTest"
-      ]
-    },
-    "depends": [
-      "quilt_loader",
-      "quilt_material_rule"
-    ]
-  }
-}
