+++ library/worldgen/surface_rule/src/main/resources/assets.quilt_surface_rule/icon.png
