+++ library/worldgen/surface_rule/src/main/resources/quilt_surface_rule.mixins.json
@@ -1,0 +1,13 @@
+{
+  "required": true,
+  "package": "org.quiltmc.qsl.worldgen.surface_rule.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "BootstrapMixin",
+    "SequenceMaterialRuleMixin",
+    "VanillaSurfaceRulesMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
