+++ library/worldgen/surface_rule/src/testmod/resources/quilt.mod.json
@@ -1,0 +1,23 @@
+{
+  "schema_version": 1,
+  "quilt_loader": {
+    "group": "org.quiltmc.qsl.worldgen",
+    "id": "quilt_surface_rule_testmod",
+    "version": "1.0.0",
+    "metadata": {
+      "name": "Quilt Surface Rule API Test Mod",
+      "license": "Apache-2.0"
+    },
+    "intermediate_mappings": "net.fabricmc:intermediary",
+    "load_type": "always",
+    "entrypoints": {
+      "events": [
+        "org.quiltmc.qsl.worldgen.surface_rule.test.QuiltSurfaceRuleTest"
+      ]
+    },
+    "depends": [
+      "quilt_loader",
+      "quilt_surface_rule"
+    ]
+  }
+}
